move effects into a single restriction file
