+++ canvas_sdk/effects/note/__init__.py
@@ -6,8 +6,10 @@ from canvas_sdk.effects.note.appointment import (
 )
 from canvas_sdk.effects.note.base import AppointmentIdentifier
 from canvas_sdk.effects.note.note import Note
-from canvas_sdk.effects.note.note_restrictions import NoteRestrictionsEffect
-from canvas_sdk.effects.note.restrictions_updated import NoteRestrictionsUpdatedEffect
+from canvas_sdk.effects.note.restrictions import (
+    NoteRestrictionsEffect,
+    NoteRestrictionsUpdatedEffect,
+)
 
 __all__ = __exports__ = (
     "AppointmentIdentifier",
