+++ canvas_sdk/effects/note/restrictions.py
@@ -30,4 +30,25 @@ class NoteRestrictionsEffect(_BaseEffect):
         }
 
 
-__exports__ = ("NoteRestrictionsEffect",)
+class NoteRestrictionsUpdatedEffect(_BaseEffect):
+    """
+    Signal to the frontend that note restrictions have changed and should be refetched.
+
+    Return this from a handler after any action that affects the output of
+    ``NoteRestrictionsEffect`` (e.g. writing a restriction key to NoteMetadata).
+    home-app will broadcast a WebSocket notification so every connected client
+    viewing that note refetches its restrictions in real time.
+    """
+
+    class Meta:
+        effect_type = EffectType.NOTE_RESTRICTIONS_UPDATED
+
+    note_id: str
+
+    @property
+    def values(self) -> dict[str, Any]:
+        """Values for the effect."""
+        return {"note_id": self.note_id}
+
+
+__exports__ = ("NoteRestrictionsEffect", "NoteRestrictionsUpdatedEffect")
